+++ build_js_dev.sh
@@ -4,6 +4,5 @@ source ./utils.sh
 
 version=$(get_version)
 printf "Current frontend version: %s" "$version"
-printf "\n\nRun frontend locally in watch mode:\n\n"
 
-caddy run -w
+build_js_dev_and_copy "$version"